requirements.txt
BeautifulSoup==3.2.1
Django==1.5.1
PIL==1.1.7
django-appconf==0.6
django-blog-zinnia==0.12.3
django-imagekit==2.0.4
django-mptt==0.5.5
django-tagging==0.3.1
django-transmeta==0.6.10
django-xmlrpc==0.1.4
pyparsing==1.5.7
pytz==2013b
six==1.3.0
wsgiref==0.1.2
